+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - wfsSearch: reset dependent fields at changed selection.
 - Issue #1310: legend images scale correctly now.
 - Issue #1307: print: add check if dpiForPdf is in dpiList and fall back to first list item if needed.
+- Issue #1323: draw_old - show icon id instead of translationkey when no translation was found.
 
 ---